refactor: consolidate category and city selectors into a single custom dropdown selector component
